refactor: 3-Layered-Architecture에 맞게 패키지 분리
