+++ src/main/java/roomescape/presentation/dto/ReservationResponse.java
@@ -1,8 +1,8 @@
-package roomescape.dto;
+package roomescape.presentation.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.business.domain.Reservation;
 
 public record ReservationResponse(
         Long id, String name, LocalDate date, LocalTime time
