+++ src/main/java/roomescape/presentation/dto/TimeResponse.java
@@ -1,7 +1,7 @@
-package roomescape.dto;
+package roomescape.presentation.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.Time;
+import roomescape.business.domain.Time;
 
 public record TimeResponse(Long id, LocalTime startAt) {
